interface/requirements.txt
python-docx==0.8.11
django==3.2.5
pandas==1.2.4
